Move undo stuffs out

+++ src/utils/ui-handlers/confirmDialogHandler.js
@@ -1,10 +1,9 @@
 import inboxObj from "../../component/Default Project/inbox"
-import createUndoBox from "../../component/Layout/createUndoBox"
+import executeWithAnimation from "../common/executeWithAnimation"
 import renderLayout from "../render"
 import { setCurrentState } from "../state"
-import { undoDeleteHandler } from "./undoHandler"
+import { undoDelete } from "./undoHandler"
 
-let undoBoxQueue = [] // Queue to store undoBox elements
 
 export function deleteConfirmHandler(project, index){
     const confirmDialog = document.getElementById('confirm__dialog')
@@ -12,22 +11,19 @@ export function deleteConfirmHandler(project, index){
     const noBtn = confirmDialog.querySelector('.close__btn')
 
     yesBtn.addEventListener('click', () => {
-        const deleteItem = project.getItem(index)
-        project.deleteItem(index)
-        confirmDialog.close('deleted')
-        renderLayout()
-        setTimeout(() => {
-            const content = document.querySelector('.content')
-            const undoBox = createUndoBox('1 item deleted')
-            content.appendChild(undoBox)
-            undoDeleteHandler(project, deleteItem, undoBox) // Pass undoBox to undoDeleteHandler
-            addToQueue(undoBox) // Add undoBox to the queue
-            checkUndoBoxQueue() // Check the queue for handling undoBox elements
-        }, 500)
+        executeWithAnimation(confirmDialog, () => {
+            const deleteItem = project.getItem(index)
+            project.deleteItem(index)
+            confirmDialog.close('deleted')
+            renderLayout()
+            undoDelete(project,deleteItem)
+        })
     })
 
     noBtn.addEventListener('click', () => {
-        confirmDialog.close('cancelled')
+        executeWithAnimation(confirmDialog, () => {
+            confirmDialog.close('cancelled')
+        })
     })
 }
 
@@ -37,44 +33,20 @@ export function deleteProjectConfirmHandler(project, index){
     const noBtn = confirmDialog.querySelector('.close__btn')
 
     yesBtn.addEventListener('click', () => {
-        const deleteItem = project.getItem(index)
-        project.deleteItem(index)
-        confirmDialog.close('deleted')
-        setCurrentState(inboxObj)
-        renderLayout()
-        setTimeout(() => {
-            const content = document.querySelector('.content')
-            const undoBox = createUndoBox('1 item deleted')
-            content.appendChild(undoBox)
-            undoDeleteHandler(project, deleteItem, undoBox) // Pass undoBox to undoDeleteHandler
-            addToQueue(undoBox) // Add undoBox to the queue
-            checkUndoBoxQueue() // Check the queue for handling undoBox elements
-        }, 500)
+        executeWithAnimation(confirmDialog, () => {
+            const deleteItem = project.getItem(index)
+            project.deleteItem(index)
+            confirmDialog.close('deleted')
+            setCurrentState(inboxObj)
+            renderLayout()
+            undoDelete(project,deleteItem)
+        })
     })
 
     noBtn.addEventListener('click', () => {
-        confirmDialog.close('cancelled')
+        executeWithAnimation(confirmDialog, () => {
+            confirmDialog.close('cancelled')
+        })
     })
 }
 
-export function addToQueue(item){
-    undoBoxQueue.push(item)
-} 
-
-export function removeUndoBox(undoBox) {
-    const index = undoBoxQueue.indexOf(undoBox)
-    if (index !== -1) {
-        undoBoxQueue.splice(index, 1)
-        undoBox.parentElement.removeChild(undoBox) // Remove the parent element containing undoBox
-    }
-}
-
-export function checkUndoBoxQueue() {
-    setTimeout(() => {
-        const undoBox = undoBoxQueue.shift() // Get the first undoBox from the queue
-        if (undoBox) {
-            undoBox.parentElement.removeChild(undoBox) // Remove the undoBox from the DOM
-            checkUndoBoxQueue() // Check the queue for more undoBox elements
-        }
-    }, 5000)
-}
